Refactor code style
