+++ src/main/java/online/carsharing/dto/response/CarResponseDto.java
@@ -2,13 +2,13 @@ package online.carsharing.dto.response;
 
 import java.math.BigDecimal;
 import lombok.Data;
-import online.carsharing.entity.Car;
+import online.carsharing.entity.Type;
 
 @Data
 public class CarResponseDto {
     private Long id;
     private String model;
     private String brand;
-    private Car.Type type;
+    private Type type;
     private BigDecimal dailyFee;
 }
