+++ src/main/java/online/carsharing/entity/Type.java
@@ -1,0 +1,8 @@
+package online.carsharing.entity;
+
+public enum Type {
+    SEDAN,
+    SUV,
+    HATCHBACK,
+    UNIVERSAL
+}
