+++ src/main/java/online/carsharing/mapper/CarMapper.java
@@ -1,7 +1,7 @@
 package online.carsharing.mapper;
 
 import online.carsharing.config.MapperConfig;
-import online.carsharing.dto.request.CarRequestDto;
+import online.carsharing.dto.request.CreateCarRequestDto;
 import online.carsharing.dto.response.CarResponseDto;
 import online.carsharing.entity.Car;
 import org.mapstruct.Mapper;
@@ -10,5 +10,5 @@ import org.mapstruct.Mapper;
 public interface CarMapper {
     CarResponseDto toDto(Car car);
 
-    Car toCar(CarRequestDto requestDto);
+    Car toCar(CreateCarRequestDto requestDto);
 }
