+++ src/main/java/online/carsharing/repository/car/CarRepository.java
@@ -2,8 +2,6 @@ package online.carsharing.repository.car;
 
 import online.carsharing.entity.Car;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
-@Repository
 public interface CarRepository extends JpaRepository<Car, Long> {
 }
