+++ src/main/java/online/carsharing/service/CarService.java
@@ -1,13 +1,13 @@
 package online.carsharing.service;
 
 import java.util.List;
-import online.carsharing.dto.request.CarRequestDto;
+import online.carsharing.dto.request.CreateCarRequestDto;
 import online.carsharing.dto.response.CarResponseDto;
 import online.carsharing.dto.update.CarUpdateDto;
 import org.springframework.data.domain.Pageable;
 
 public interface CarService {
-    CarResponseDto save(CarRequestDto newCar);
+    CarResponseDto save(CreateCarRequestDto newCar);
 
     List<CarResponseDto> findAll(Pageable pageable);
