+++ src/main/resources/db/changelog/changes/chopyk/create-cars-table.yaml
@@ -12,7 +12,6 @@ databaseChangeLog:
                   autoIncrement: true
                   constraints:
                     primaryKey: true
-                    nullable: false
               - column:
                   name: model
                   type: varchar(255)
@@ -25,7 +24,7 @@ databaseChangeLog:
                     nullable: false
               - column:
                   name: type
-                  type: enum('SEDAN', 'SUV', 'HATCHBACK', 'UNIVERSAL')
+                  type: varchar(255)
                   constraints:
                     nullable: false
               - column:
@@ -39,9 +38,8 @@ databaseChangeLog:
                   constraints:
                     nullable: false
               - column:
-                  name: is_deleted
+                  name: deleted
                   type: boolean
                   defaultValueBoolean: false
                   constraints:
                     nullable: false
-
