+++ src/main/resources/db/changelog/db.changelog-master.yaml
@@ -1,3 +1,3 @@
 databaseChangeLog:
   - include:
-      file: /db/changelog/changes/create-cars-table.yml
+      file: /db/changelog/changes/chopyk/create-cars-table.yaml
