+++ src/test/resources/application.properties
@@ -1,11 +1,5 @@
-spring.application.name=car-sharing-service
-spring.datasource.url=jdbc:mysql://localhost:3306/car_sharing
-spring.datasource.username=root
+spring.datasource.url=jdbc:h2:mem:testdb
+spring.datasource.driverClassName=org.h2.Driver
+spring.datasource.username=sa
 spring.datasource.password=password
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-
-spring.jpa.hibernate.ddl-auto=validate
-spring.jpa.show-sql=true
-spring.jpa.open-in-view=false
-
-server.servlet.context-path=/api
+spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
